+++ Changelog.md
@@ -1,6 +1,7 @@
 (Please keep in mind that these are only the major/important changes. There have been way more than written here)<br>
 
 # 0.6.1.1
+- Changed guild channel creation stuff (sorry for name changes).
 - Re-added GetInvite() method.
 - Added some OAuth2 support.
 - Improved Connections functionality.